requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.2.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
pandas>=2.0.0
tabulate>=0.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0